Fix lexing of Decimal and Identifier
The lexer was mistaking 'e' in an Identifier to mean scientific
notation. By iterating over every character in the lexeme, we can
subvert this issue. This will make the lexer slower, but hopefully
that will not be an issue later down the road. There is probably a
more elegant way to do this.