Simplify lexing of Identifiers
The previous implementation worked, but it was slow, and the code was
messy and confusing. This commit does away with the []bool, which
checked the current lexeme for alphanumeric characters do determine if
it was an Identifier or a Decimal. Instead, we range over the lexeme
and check if any of the characters are alphanumeric. If they are, it's
an Identifier. If not, it's a Decimal.